#20 stats page has been given a 21st century makeover, complete with Chart.js implementation
