+++ css/styles.css
@@ -336,3 +336,8 @@ html, body {
 {
   color: black;
 }
+
+.stats-cell
+{
+  text-align: center;
+}
